+++ taxsystem/auth_hooks.py
@@ -1,12 +1,13 @@
 """Hook into Alliance Auth"""
 
 # Django
-# Alliance Auth
 from django.utils.translation import gettext_lazy as _
 
+# Alliance Auth
 from allianceauth import hooks
 from allianceauth.services.hooks import MenuItemHook, UrlHook
 
+# AA TaxSystem
 from taxsystem.models.filters import FilterAmount, FilterReason
 
 from . import app_settings, urls