+++ taxsystem/tests/test_view.py
@@ -1,12 +1,16 @@
 """Tax System for a TestView class."""
 
+# Standard Library
 from http import HTTPStatus
 
+# Django
 from django.test import RequestFactory, TestCase
 from django.urls import reverse
 
+# Alliance Auth (External Libs)
 from app_utils.testdata_factories import UserMainFactory
 
+# AA TaxSystem
 from taxsystem.views import index