+++ taxsystem/urls.py
@@ -1,7 +1,9 @@
 """App URLs"""
 
+# Django
 from django.urls import path, re_path
 
+# AA TaxSystem
 from taxsystem import views
 from taxsystem.api import api
 
@@ -10,6 +12,7 @@ app_name: str = "taxsystem"
 urlpatterns = [
     # -- Tax System
     path("", views.index, name="index"),
+    path("admin/", views.admin, name="admin"),
     path(
         "corporation/<int:corporation_id>/view/payments/",
         views.payments,